Many improvements
- Fix AirTime calculation
- Enum classes as uint8_t type
- Reduce RAM usage
- Remove virtual destructors
- Add minValue and maxValue to ConfigItem
- Limit humidity to 0-100 after compensation
- Add define DEBUG_LORA_MESSAGE
- Temporary disable check of msg id in LoRa header
- Remove len byte in LoRa header
- Increase rssi to two bytes
- Remove entity id in value request
- Insert delay between sending value messages

+++ AirTime.h
@@ -7,7 +7,8 @@
 constexpr uint8_t SECONDS_PER_MINUTE = 60;
 constexpr uint8_t MINUTES_PER_HOUR = 60;
 constexpr uint16_t MS_PER_MINUTE = SECONDS_PER_MINUTE * 1000;
-constexpr uint32_t MS_PER_HOUR = MINUTES_PER_HOUR * MS_PER_MINUTE;
+constexpr uint32_t MS_PER_HOUR =
+    static_cast<uint32_t>(MINUTES_PER_HOUR) * MS_PER_MINUTE;
 
 class AirTime {
  public:
